requirements.txt
Django>=3,<4
djangorestframework>=3.14,<3.15
drf-spectacular>=0.26.1,<0.27
psycopg2>=2.9,<2.10